separate inline JS tests from JS parsing tests
