+++ test/spec/JSUtils-test-files/eof.js
