+++ test/spec/JSUtils-test-files/eof2.js
