+++ test/spec/JSUtils-test-files/jquery-1.7.js
