+++ test/spec/JSUtils-test-files/simple.js
